Crate megstd

MEG-OS standard library like std

Re-exports

pub use meggl as drawing;

pub use crate::osstr::*;

Modules

error

fs

game

io

mem

osstr

path

rand

string

sys

time

uuid

Macros

format

Creates a String using interpolation of runtime expressions.

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BTreeMap

An ordered map based on a B-Tree.

Box

A pointer type that uniquely owns a heap allocation of type T.

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Weak

Weak is a version of Arc that holds a non-owning reference to the managed allocation. The allocation is accessed by calling upgrade on the Weak pointer, which returns an [Option]<Arc<T>>.

Traits

ToOwned

A generalization of Clone to borrowed data.

ToString

A trait for converting a value to a String.