Enum megstd::sys::megos::svc::Function

#[repr(u32)]
pub enum Function {

    Exit = 0,
    PrintString = 1,
    Monotonic = 2,
    Time = 3,
    Usleep = 4,
    GetSystemInfo = 5,
    NewWindow = 6,
    CloseWindow = 7,
    BeginDraw = 8,
    EndDraw = 9,
    DrawString = 10,
    FillRect = 11,
    DrawRect = 12,
    DrawLine = 13,
    Blt8 = 14,
    Blt1 = 15,
    Blt32 = 16,
    BlendRect = 17,
    WaitChar = 18,
    ReadChar = 19,
    DrawShape = 20,
    Rand = 100,
    Srand = 101,
    Alloc = 102,
    Dealloc = 103,
    MkDir = 104,
    RmDir = 105,
    ChDir = 106,
    Open = 107,
    Close = 108,
    Read = 109,
    Write = 110,
    LSeek = 111,
    IoCtl = 112,
    Unlink = 113,
    OpenDir = 114,
    ReadDir = 115,

}

MEG-OS Maystorm System Call Function Numbers

Variants

Exit = 0

Terminate the process normally

PrintString = 1

Display a string

Monotonic = 2

Get the value of the monotonic timer in microseconds

Time = 3

Perform the time service

Usleep = 4

Blocks a thread for the specified microseconds

GetSystemInfo = 5

Get the system information

NewWindow = 6

Create a new window

CloseWindow = 7

Close a window

BeginDraw = 8

Create a drawing context

EndDraw = 9

Discard the drawing context and reflect it to the screen

DrawString = 10

Draw a string in a window

FillRect = 11

Fill a rectangle in a window

DrawRect = 12

Draw a rectangle in a window

DrawLine = 13

Draw a line in a window

Blt8 = 14

Draw a bitmap in a window

Blt1 = 15

Draw a bitmap in a window

Blt32 = 16

Draw a bitmap in a window

BlendRect = 17

Blend (test)

WaitChar = 18

Wait for char event

ReadChar = 19

Read a char event

DrawShape = 20

Draw a shape in a window

Rand = 100

Returns a simple pseudo-random number

Srand = 101

Set the seed of the random number

Alloc = 102

Allocates memory blocks with a simple allocator

Dealloc = 103

Frees an allocated memory block

MkDir = 104

RmDir = 105

ChDir = 106

Open = 107

Close = 108

Read = 109

Write = 110

LSeek = 111

IoCtl = 112

Unlink = 113

OpenDir = 114

ReadDir = 115